use aws secrets manager in a reusable way
