+++ sites/zappa/requirements.txt
@@ -2,3 +2,4 @@ flask >= 1.1.1
 zappa >= 0.48
 flask-dance >= 3.0.0
 flask-markdown >= 0.3
+botocore >= 1.10.3
